requirements.txt
langchain
langchain_core
langchain_community
langchain-groq
langchain-ollama
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
streamlit